overwrite old encode.py with detection_encoder.py, which now does it all
